% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_based_calculations.R
\name{area_based_calculations}
\alias{area_based_calculations}
\title{Calculate the population results for the area based methods}
\usage{
area_based_calculations(availUseData, sampleGroups, optionsList)
}
\arguments{
\item{availUseData}{comes from area_based_extraction}

\item{sampleGroups}{list if IDs for the sample groups}

\item{optionsList}{options list that will be used to loop through options}
}
\value{
Population estimates for area based methods.
}
\description{
A
}
