% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_based_extraction.R
\name{area_based_extraction}
\alias{area_based_extraction}
\title{Extract all required values for area methods}
\usage{
area_based_extraction(allIndividualData, optionsList)
}
\arguments{
\item{allIndividualData}{The full list of all simulated individuals, subsampled into regimes}

\item{optionsList}{}
}
\value{
a
}
\description{
A
}
