% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_available_area.R
\name{build_available_area}
\alias{build_available_area}
\title{Generate the prerequisite for building a polygon for use as an available area}
\usage{
build_available_area(
  movementData,
  method = c("MCP", "KDElscv", "KDEhref", "AKDE", "dBBMM"),
  SRS_string = "EPSG:32601",
  dBBMMsettings = NULL
)
}
\arguments{
\item{movementData}{must have a x and y column for locations, and a datetime
column for timestamps ("\%Y-\%m-\%d \%H:\%M:\%S")}

\item{method}{"MCP", "KDE_LSCV", "KDE_href", "AKDE", "dBBMM"}

\item{SRS_string}{= "EPSG:32601"}

\item{dBBMMsettings}{time duration of window and margin in hours, function
translate duration to datapoints}
}
\value{
An item that is the fit or UD that the a polygon can be
extracted from using build_available_polygon
}
\description{
A
}
