% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_available_polygon.R
\name{build_available_polygon}
\alias{build_available_polygon}
\title{Generate a polygon for use as an available area}
\usage{
build_available_polygon(
  areaResource,
  method = c("MCP", "KDElscv", "KDEhref", "AKDE", "dBBMM"),
  contour,
  SRS_string = "EPSG:32601"
)
}
\arguments{
\item{areaResource}{an output from build_available_areas that is used to
extract a polygon.}

\item{method}{"MCP", "KDE_LSCV", "KDE_href", "AKDE", "dBBMM"}

\item{contour}{Numeric, > 0 and < 100. Can be a vector.}

\item{SRS_string}{= "EPSG:32601"}
}
\value{
A polygon.
}
\description{
A
}
