% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_spec_curves.R
\name{generate_spec_curves}
\alias{generate_spec_curves}
\title{Generate specification curves}
\usage{
generate_spec_curves(outputResults, method)
}
\arguments{
\item{outputResults}{}

\item{method}{c("ssf", "area", "pois")}
}
\value{
a
}
\description{
A
}
