% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_marginals.R
\name{inla_mmarginal}
\alias{inla_mmarginal}
\title{Extract marginals from INLA}
\usage{
inla_mmarginal(inlaOutput)
}
\arguments{
\item{inlaOutput}{The output from an INLA model.}
}
\value{
abc
}
\description{
abc.
}
