% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_pois_inla.R
\name{method_pois_inla}
\alias{method_pois_inla}
\title{Calculate population selection using a Poisson model}
\usage{
method_pois_inla(allIndividualData, sampleGroups, optionsList)
}
\arguments{
\item{allIndividualData}{All simulated movement data}

\item{sampleGroups}{list if IDs for the sample groups}

\item{optionsList}{options list that will be used to loop through options}
}
\value{
Population estimates using a Poisson model.
}
\description{
A
}
