% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_ssf_results.R
\name{sample_ssf_results}
\alias{sample_ssf_results}
\title{Sample individual SSF models}
\usage{
sample_ssf_results(ssfResults, sampleGroups, optionsList)
}
\arguments{
\item{ssfResults}{all sff models and corresponding options data.frame}

\item{sampleGroups}{list if IDs for the sample groups}
}
\value{
Population estimates for SSF based methods.
}
\description{
create a population estimate
}
