% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_data.R
\name{subset_frequency}
\alias{subset_frequency}
\title{Subset simulated tracking data}
\usage{
subset_frequency(movementData, freqPreset)
}
\arguments{
\item{movementData}{The locations data.frame from the abmAnimalMovement
simulation.}

\item{freqPreset}{Must be one of the following: 0.5, 1.0, 2.0, 6.0, 12.0,
24.0, 48.0, 168.0 hours}

\item{daysDuration}{Duration in days}
}
\value{
A data.frame reduced down to desired duration and frequency preset.
}
\description{
A function to subset simulated tracked data, either by frequency
or by duration. The input data.frame must have a POSIXct datetime column.
}
