% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_indi_ssf.R
\name{wrapper_indi_ssf}
\alias{wrapper_indi_ssf}
\title{Run all ssf options}
\usage{
wrapper_indi_ssf(allIndividualData, optionsList)
}
\arguments{
\item{optionsList}{Must have the following items: Method_method,
MethodSSF_mf, MethodSSF_sd, MethodSSF_td, MethodSSF_as}

\item{movementData}{must have a x and y column for locations, and a datetime
column for timestamps ("\%Y-\%m-\%d \%H:\%M:\%S")}
}
\value{
a
}
\description{
A
}
